Add the ability to boot without dhcp
The primary purpose of this is to experiment with KernelCommandLine
constraints.  Eventually this mechanism will probably be used to
allow static configuration to be specified on the command line or
networking.json.  Until Triton integration comes along further, it
is probably useful for standalone boots that get delayed by lack of
dhcp responses.